% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.is.R
\name{as.is}
\alias{as.is}
\title{Title}
\usage{
as.is(x, ...)
}
\arguments{
\item{x}{factor or charactor}
}
\value{
x factor
}
\description{
Title
}
\examples{
(a <- c("character", "your", "eyeballs","hits","hits"))
(a <- factor(c("character", "your", "eyeballs","hits","hits")))

a <- as.is(a)
a
}

